requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
anthropic
python-dotenv